+++ OKX.Api/Trade/Requests/OkxTradeOrderPlaceRequestAttachedAlgo.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// OKX Order Algo Request
 /// </summary>
-public record OkxTradeOrderPlaceAttachedAlgoRequest
+public record OkxTradeOrderPlaceRequestAttachedAlgo
 {
     /// <summary>
     /// Client-supplied Algo ID when placing order attaching TP/SL
@@ -18,7 +18,18 @@ public record OkxTradeOrderPlaceAttachedAlgoRequest
     /// For condition TP order, if you fill in this parameter, you should fill in the take-profit order price as well.
     /// </summary>
     [JsonProperty("tpTriggerPx", NullValueHandling = NullValueHandling.Ignore)]
-    public string? TakeProfitTriggerPrice { get; set; }
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
+    public decimal? TakeProfitTriggerPrice { get; set; }
+
+    /// <summary>
+    /// Take profit trigger ratio, 0.3 represents 30%
+    /// Only one of tpTriggerPx and tpTriggerRatio can be passed
+    /// Only applicable to FUTURES and SWAP.
+    /// If the main order is a buy order, it must be greater than 0, and if the main order is a sell order, it must be bewteen -1 and 0.
+    /// </summary>
+    [JsonProperty("tpTriggerRatio", NullValueHandling = NullValueHandling.Ignore)]
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
+    public decimal? TakeProfitTriggerRatio { get; set; }
 
     /// <summary>
     /// Take-profit order price
@@ -28,7 +39,8 @@ public record OkxTradeOrderPlaceAttachedAlgoRequest
     /// If the price is -1, take-profit will be executed at the market price.
     /// </summary>
     [JsonProperty("tpOrdPx", NullValueHandling = NullValueHandling.Ignore)]
-    public string? TakeProfitOrderPrice { get; set; }
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
+    public decimal? TakeProfitOrderPrice { get; set; }
 
     /// <summary>
     /// TP order kind
@@ -37,14 +49,25 @@ public record OkxTradeOrderPlaceAttachedAlgoRequest
     /// The default is condition
     /// </summary>
     [JsonProperty("tpOrdKind", NullValueHandling = NullValueHandling.Ignore)]
-    public string TakeProfitOrderKind { get; set; } = "condition";
+    public OkxAlgoOrderKind? TakeProfitOrderKind { get; set; } = OkxAlgoOrderKind.Condition;
 
     /// <summary>
     /// Stop-loss trigger price
     /// If you fill in this parameter, you should fill in the stop-loss order price.
     /// </summary>
     [JsonProperty("slTriggerPx", NullValueHandling = NullValueHandling.Ignore)]
-    public string? StopLossTriggerPrice { get; set; }
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
+    public decimal? StopLossTriggerPrice { get; set; }
+
+    /// <summary>
+    /// Stop profit trigger ratio, 0.3 represents 30%
+    /// Only one of slTriggerPx and slTriggerRatio can be passed
+    /// Only applicable to FUTURES and SWAP.
+    /// If the main order is a buy order, it should be bewteen 0 and 1, and if the main order is a sell order, it must be greater than 0.
+    /// </summary>
+    [JsonProperty("slTriggerRatio", NullValueHandling = NullValueHandling.Ignore)]
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
+    public decimal? StopLossTriggerRatio { get; set; }
 
     /// <summary>
     /// Stop-loss order price
@@ -52,6 +75,7 @@ public record OkxTradeOrderPlaceAttachedAlgoRequest
     /// If the price is -1, stop-loss will be executed at the market price.
     /// </summary>
     [JsonProperty("slOrdPx", NullValueHandling = NullValueHandling.Ignore)]
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
     public string? StopLossOrderPrice { get; set; }
 
     /// <summary>
@@ -70,7 +94,8 @@ public record OkxTradeOrderPlaceAttachedAlgoRequest
     /// Size. Only applicable to TP order of split TPs, and it is required for TP order of split TPs
     /// </summary>
     [JsonProperty("sz", NullValueHandling = NullValueHandling.Ignore)]
-    public string? Size { get; set; }
+    [JsonConverter(typeof(DecimalAsStringNullableConverter))]
+    public decimal? Size { get; set; }
 
     /// <summary>
     /// Whether to enable Cost-price SL. Only applicable to SL order of split TPs. Whether slTriggerPx will move to avgPx when the first TP order is triggered
@@ -78,5 +103,6 @@ public record OkxTradeOrderPlaceAttachedAlgoRequest
     /// 1: Enable
     /// </summary>
     [JsonProperty("amendPxOnTriggerType", NullValueHandling = NullValueHandling.Ignore)]
-    public string AmendPriceOnTriggerType { get; set; } = "0";
+    [JsonConverter(typeof(DecimalAsStringNullableConverter), "1", "0")]
+    public bool AmendPriceOnTriggerType { get; set; }
 }